refactor(index): break on-disk cache out to its own module
